Gen-AI-Projects/requirements.txt
streamlit
pypdf
openai
sentence-transformers
faiss-cpu
langchain
langchain-huggingface
langchain-text-splitters
langchain-community
langchain-openai
python-dotenv
unstructured
tiktoken
pinecone-client
langchain-pinecone
pandas
numpy 
ipykernel
ctransformers
fastapi
python-box
google-generativeai

